pyproject.toml
[project]
name = "auto-ppt-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "langchain-core>=1.2.18",
    "langchain[openai]>=1.2.10",
    "langgraph>=1.0.10",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]
